quick fix of tests: new version file, merged old Conf class with new CommandLineOptions
TODO: read Configuration once when creating CommandLineOptions from input file
