--- src/main/resources/Version.properties
@@ -1,1 +1,0 @@
-application.version=${project.version}
